Tfb101/update reports fix cm abundance (#182)
* fix conf_mat.scale_by_abundance for float abundances, add imposter explore cell to classify report, minor other report changes

* add zest for conf_mat scale_by_abundance to ensure correct float scaling

* merge master

+++ plaster/run/call_bag.py
@@ -764,7 +764,7 @@ class CallBag:
         return s_at_prec
 
     def conf_mat_at_score_threshold(self, score_thresh):
-        return self.conf_mat(mask=self.scores >= score_thresh.astype(self.scores.dtype))
+        return self.conf_mat(mask=self.scores >= score_thresh)
 
     def false_rates_by_pep(self, pep_i, at_prec, n_false):
         """
